+++ src/fennflow/backends/sqlalchemy/_query_flows/is_empty.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 from dataclasses import dataclass
 
 from fennflow._query_specs.select.is_empty import IsEmptyQuerySpec
-from fennflow.backends.sqlalchemy import exists, select
 from fennflow.backends.sqlalchemy._query_flows.base import (
     BaseSqlalchemyBackendQueryFlow,
 )
@@ -15,6 +14,8 @@ class IsEmptyFlow(BaseSqlalchemyBackendQueryFlow[IsEmptyQuerySpec, bool]):
         self,
         query_spec: IsEmptyQuerySpec,
     ) -> bool:
+        from fennflow.backends.sqlalchemy._base import exists, select
+
         model = self.adapter.orm_model
         stmt = select(~exists().where(model.scope == query_spec.scope))
         result = await self.session.execute(stmt)
