+++ src/fennflow/backends/sqlalchemy/_query_flows/merge/core.py
@@ -17,7 +17,6 @@ MergeFlowStrategy = Callable[
     Awaitable[None],
 ]
 
-
 fallback_registry = FallbackRegistry[
     Dialect | str,
     MergeFlowStrategy,
