+++ src/fennflow/backends/sqlalchemy/_query_flows/merge/postgres.py
@@ -2,8 +2,6 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from sqlalchemy.dialects.postgresql import insert as pg_insert
-
 from fennflow._shared import unique_constraint
 from fennflow.backends.sqlalchemy._query_flows.utils.postgres import get_update_fields
 
@@ -16,6 +14,8 @@ async def run(
     flow: MergeFlow,
     query_spec: InsertQuerySpec,
 ) -> None:
+    from fennflow.backends.sqlalchemy._base import pg_insert
+
     rows = tuple(flow.adapter.to_orm(r).model_dump() for r in query_spec.records)
 
     model = flow.adapter.orm_model
