+++ src/fennflow/backends/sqlalchemy/_query_flows/utils/postgres.py
@@ -7,13 +7,13 @@ from fennflow._shared import unique_constraint
 if TYPE_CHECKING:
     from sqlalchemy.dialects.postgresql import Insert
 
-    from fennflow.backends.sqlalchemy import AbstractOperationRecordModel
+    from fennflow.backends.sqlalchemy._base import AbstractOperationRecordModel
 
 
 def get_update_fields(
-    insert_stmt: Insert,
-    model: type[AbstractOperationRecordModel],
-) -> dict:
+        insert_stmt: Insert,
+        model: type[AbstractOperationRecordModel],
+        ) -> dict:
     """Returns fields required to update an operation.
 
         **Example**::
@@ -42,4 +42,4 @@ def get_update_fields(
         col.name: insert_stmt.excluded[col.name]
         for col in model.__table__.columns
         if col.name not in unique_constraint and not col.primary_key
-    }
+        }
